Got rid of file_size_t, made offset_t unsigned (now used in place of file_size_t), and added rel_offset_t (for cases where negative offsets are required, such as seeking on a file handle).

+++ source/drivers/disk/partition.c
@@ -38,7 +38,7 @@ static int partition_device_read(device_t *device, void *data, void *buf, size_t
 	partition_t *part = device->data;
 	disk_device_t *disk = part->parent;
 
-	if((uint64_t)(offset + count) >= (part->size * disk->blksize)) {
+	if((offset + count) >= (part->size * disk->blksize)) {
 		*bytesp = 0;
 		return 0;
 	}
@@ -59,7 +59,7 @@ static int partition_device_write(device_t *device, void *data, const void *buf,
 	partition_t *part = device->data;
 	disk_device_t *disk = part->parent;
 
-	if((uint64_t)(offset + count) >= (part->size * disk->blksize)) {
+	if((offset + count) >= (part->size * disk->blksize)) {
 		*bytesp = 0;
 		return 0;
 	}
